requirements.txt
groq
phidata 
yfinance
duckduckgo-search
